# docs/openapi.yaml
openapi: 3.1.0
info:
  title: VelesDB API
  description: High-performance vector database for AI applications. Supports semantic search, HNSW indexing, and multiple distance metrics. Authentication is optional — when API keys are configured via VELESDB_API_KEYS, all endpoints except /health and /ready require a valid Bearer token.
  contact:
    name: VelesDB Team
    url: https://github.com/cyberlife-coder/VelesDB
  license:
    name: VelesDB Core License 1.0
    url: https://github.com/cyberlife-coder/VelesDB/blob/main/LICENSE
  version: 6.0.0
servers:
- url: /
  description: Local server
paths:
...
  /collections/{name}/search:
    post:
      tags:
      - search
      summary: Search for similar vectors.
      description: |-
        Auto-detects search mode:
        - **Dense**: `vector` only (existing behavior)
        - **Sparse**: `sparse_vector` only
        - **Hybrid**: both `vector` and `sparse_vector` (fused via RRF/RSF)
      operationId: search
      parameters:
      - name: name
        in: path
        description: Collection name
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
        required: true
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /collections/{name}/search/batch:
    post:
      tags:
      - search
      summary: Batch search for multiple vectors.
      operationId: batch_search
      parameters:
      - name: name
        in: path
        description: Collection name
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchSearchRequest'
        required: true
      responses:
        '200':
          description: Batch search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchSearchResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /collections/{name}/search/hybrid:
    post:
      tags:
      - search
      summary: Hybrid search combining vector similarity and BM25 text search.
      operationId: hybrid_search
      parameters:
      - name: name
        in: path
        description: Collection name
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HybridSearchRequest'
        required: true
      responses:
        '200':
          description: Hybrid search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /collections/{name}/search/ids:
    post:
      tags:
      - search
      summary: Lightweight search returning only IDs and scores (no payload hydration).
      description: |-
        Accepts the same request body as the standard `/search` endpoint, for a
        dense, sparse or hybrid search. `mode` is checked on every shape but
        applied only to a dense search with neither a `filter` nor `ef_search`:
        `ef_search` wins over it, and a filtered search applies neither (#457).
      operationId: search_ids
      parameters:
      - name: name
        in: path
        description: Collection name
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
        required: true
      responses:
        '200':
          description: IDs-only search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchIdsResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /collections/{name}/search/multi:
    post:
      tags:
      - search
      summary: Multi-query search with fusion strategies.
      operationId: multi_query_search
      parameters:
      - name: name
        in: path
        description: Collection name
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MultiQuerySearchRequest'
        required: true
      responses:
        '200':
          description: Multi-query search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /collections/{name}/search/multi/ids:
    post:
      tags:
      - search
      summary: Multi-query fusion search returning only ids and scores (no payloads).
      description: |-
        Faster than `/search/multi` when payloads are not needed: the core
        `multi_query_search_ids` kernel skips payload hydration. Metadata filters
        are not supported here — use `/search/multi` for filtered fusion.
      operationId: multi_query_search_ids
      parameters:
      - name: name
        in: path
        description: Collection name
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MultiQuerySearchRequest'
        required: true
      responses:
        '200':
          description: Multi-query ids-only results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchIdsResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /collections/{name}/search/text:
    post:
      tags:
      - search
      summary: Search using BM25 full-text search.
      operationId: text_search
      parameters:
      - name: name
        in: path
        description: Collection name
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TextSearchRequest'
        required: true
      responses:
        '200':
          description: Text search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    BatchSearchRequest:
      type: object
      description: Request for batch vector search.
      required:
      - searches
      properties:
        searches:
          type: array
          items:
            $ref: '#/components/schemas/SearchRequest'
          description: List of search requests.
    BatchSearchResponse:
      type: object
      description: Response from batch search.
      required:
      - results
      - timing_ms
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResponse'
          description: Results for each search query.
        timing_ms:
          type: number
          format: double
          description: Total time in milliseconds.
...
    ErrorResponse:
      type: object
      description: Error response.
      required:
      - error
      properties:
        code:
          type:
          - string
          - 'null'
          description: Structured error code (e.g., `"VELES-004"` for dimension mismatch).
        error:
          type: string
          description: Error message.
...
    FusionRequest:
      type: object
      description: Fusion configuration for hybrid dense+sparse search.
      required:
      - strategy
      properties:
        avg_w:
          type:
          - number
          - 'null'
          format: float
          description: |-
            Weighted fusion: weight applied to the average score component
            (only for strategy = "weighted", default 0.5).
        dense_w:
          type:
          - number
          - 'null'
          format: float
          description: Dense weight (only for strategy = "rsf", default 0.5).
        hit_w:
          type:
          - number
          - 'null'
          format: float
          description: |-
            Weighted fusion: weight applied to the hit-ratio component
            (only for strategy = "weighted", default 0.2).
        k:
          type:
          - integer
          - 'null'
          format: int32
          description: RRF k parameter (only for strategy = "rrf", default 60).
          minimum: 0
        max_w:
          type:
          - number
          - 'null'
          format: float
          description: |-
            Weighted fusion: weight applied to the maximum score component
            (only for strategy = "weighted", default 0.3).
        sparse_w:
          type:
          - number
          - 'null'
          format: float
          description: Sparse weight (only for strategy = "rsf", default 0.5).
        strategy:
          type: string
          description: |-
            Fusion strategy: "rrf", "rsf" (alias "`relative_score`"),
            "average" (alias "avg"), "maximum" (alias "max"), or "weighted".
          example: rrf
...
    HybridSearchRequest:
      type: object
      description: Request for hybrid search (vector + text).
      required:
      - vector
      - query
      properties:
        filter:
          type: object
          description: Optional metadata filter.
          example:
            condition:
              field: category
              type: eq
              value: tech
        query:
          type: string
          description: Text query for BM25 search.
          example: rust programming
        top_k:
          type: integer
          description: Number of results to return.
          example: 10
          minimum: 0
        vector:
          type: array
          items:
            type: number
            format: float
          description: Query vector for similarity search.
        vector_weight:
          type: number
          format: float
          description: Weight for vector similarity (0.0-1.0). Text weight = 1 - `vector_weight`.
          example: 0.5
    IdScoreResult:
      type: object
      description: A single ID+score result from IDs-only search.
      required:
      - id
      - score
      properties:
        id:
          type: string
          description: Point ID.
        score:
          type: number
          format: float
          description: Similarity score.
...
    MultiQuerySearchRequest:
      type: object
      description: Request for multi-query vector search with fusion.
      required:
      - vectors
      properties:
        avg_weight:
          type: number
          format: float
          description: 'Weighted fusion: weight for average score component.'
          example: 0.5
        dense_weight:
          type: number
          format: float
          description: 'Relative score fusion: weight for the dense (vector) branch.'
          example: 0.5
        filter:
          type: object
          description: Optional metadata filter.
          example:
            condition:
              field: category
              type: eq
              value: tech
        hit_weight:
          type: number
          format: float
          description: 'Weighted fusion: weight for hit count component.'
          example: 0.2
        max_weight:
          type: number
          format: float
          description: 'Weighted fusion: weight for max score component.'
          example: 0.3
        rrf_k:
          type: integer
          format: int32
          description: RRF k parameter (only used when strategy = "rrf").
          example: 60
          minimum: 0
        sparse_weight:
          type: number
          format: float
          description: 'Relative score fusion: weight for the sparse branch.'
          example: 0.5
        strategy:
          type: string
          description: 'Fusion strategy: "average", "maximum", "rrf", "weighted", "`relative_score`".'
          example: rrf
        top_k:
          type: integer
          description: Number of results to return.
          example: 10
          minimum: 0
        vectors:
          type: array
          items:
            type: array
            items:
              type: number
              format: float
          description: List of query vectors.
...
    SearchIdsResponse:
      type: object
      description: Response from IDs-only search (no payload hydration).
      required:
      - results
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/IdScoreResult'
          description: Search results with IDs and scores only.
    SearchRequest:
      type: object
      description: Request for vector search (dense, sparse, or hybrid).
      properties:
        ef_search:
          type:
          - integer
          - 'null'
          description: |-
            HNSW `ef_search` parameter, in `[16, 4096]`. Any other value is
            refused with a `400` naming the range (#2274). A batch entry's is
            checked but not applied, like `mode`.
          example: 128
          minimum: 0
        filter:
          type: object
          description: Optional metadata filter.
          example:
            condition:
              field: category
              type: eq
              value: tech
        fusion:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/FusionRequest'
            description: Fusion configuration for hybrid search.
        mode:
          type:
          - string
          - 'null'
          description: |-
            Search quality preset, in any letter case: `fast`, `balanced`,
            `accurate`, `perfect`, `autotune` (aliases `auto_tune` and `auto`),
            `custom:<ef>` or `adaptive:<min_ef>:<max_ef>` (`min_ef` at most
            `max_ef`). Any other value is refused with a `400` naming these forms
            (#2267). It applies only to a dense search with neither a `filter` nor
            `ef_search`, and a batch entry's is checked but not applied.
          example: balanced
        sparse_index:
          type:
          - string
          - 'null'
          description: Which named sparse index to search.
        sparse_vector:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/SparseVectorInput'
            description: Single sparse query vector.
        sparse_vectors:
          type:
          - object
          - 'null'
          description: Named sparse query vectors map.
          additionalProperties:
            $ref: '#/components/schemas/SparseVectorInput'
          propertyNames:
            type: string
        timeout_ms:
          type:
          - integer
          - 'null'
          format: int64
          description: Query timeout in milliseconds.
          example: 30000
          minimum: 0
        top_k:
          type: integer
          description: Number of results to return.
          minimum: 0
        vector:
          type: array
          items:
            type: number
            format: float
          description: Query vector for dense search.
    SearchResponse:
      type: object
      description: Response from vector search.
      required:
      - results
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResultResponse'
          description: Search results.
...
    SparseVectorInput:
      oneOf:
      - type: object
        description: Canonical parallel-array format with explicit indices and values.
        required:
        - indices
        - values
        properties:
          indices:
            type: array
            items:
              type: integer
              format: int32
              minimum: 0
            description: Term/dimension indices (u32).
          values:
            type: array
            items:
              type: number
              format: float
            description: Corresponding weights.
      - type: object
        description: Dictionary format mapping string term IDs to weights.
        additionalProperties:
          type: number
          format: float
        propertyNames:
          type: string
      description: |-
        Input format for sparse vectors, supporting two JSON representations:

        - **Parallel arrays**: `{"indices": [42, 1337], "values": [0.5, 1.2]}`
        - **Dict format** (Qdrant-compatible): `{"42": 0.5, "1337": 1.2}`
...
    TextSearchRequest:
      type: object
      description: Request for BM25 text search.
      required:
      - query
      properties:
        filter:
          type: object
          description: Optional metadata filter.
          example:
            condition:
              field: category
              type: eq
              value: tech
        query:
          type: string
          description: Text query for full-text search.
          example: rust programming
        top_k:
          type: integer
          description: Number of results to return.
          example: 10
          minimum: 0
...
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
security:
- bearer_auth: []
...